Create AnimalGame.py
Python script that simulates a game where the user gets a randomly generated Emoji every time they type 'Play'. It simulates 2 payment methods to pay the game, 'cash or card' and will simulate inputting credit card info if you choose card. Sometimes the script will make it so you can only choose card, or make it so when you type 'Play' it will tell you that your funds are insufficient, and you will have to restart. The objective of the game is to try and get the Alien. If you get the Alien emoji, you will win!